fix: better organization for db seeds
