+++ prisma/dummyData.js
@@ -1,0 +1,487 @@
+/**
+* Dummy Data
+* 
+* This file contains the dummy data used for seeding the database.
+* It includes default values for societies, groups, requests,
+* users, and other related entities to ensure the application has 
+* a baseline dataset for development and testing.
+*/
+
+import {
+  ADMIN_COST_CENTER_NAME,
+  ADMIN_DEPARTMENT_NAME,
+} from './seedShared.js';
+
+export const COST_CENTERS = [
+  { cost_center_name: 'CC-001', cost_center_code: 102 },
+  { cost_center_name: 'CC-002', cost_center_code: 103 },
+  { cost_center_name: 'CC-003', cost_center_code: 104 },
+  { cost_center_name: 'CC-004', cost_center_code: 105 },
+  { cost_center_name: 'CC-005', cost_center_code: 106 },
+  { cost_center_name: 'CC-006', cost_center_code: 107 },
+  { cost_center_name: ADMIN_COST_CENTER_NAME, cost_center_code: 101 },
+];
+
+export const DEPARTMENTS = [
+  { name: 'Finanzas', costCenter: 'CC-001', active: true },
+  { name: 'Recursos Humanos', costCenter: 'CC-002', active: true },
+  { name: 'Tecnología', costCenter: 'CC-003', active: true },
+  { name: 'Marketing', costCenter: 'CC-004', active: true },
+  { name: 'Operaciones', costCenter: 'CC-005', active: true },
+  { name: ADMIN_DEPARTMENT_NAME, costCenter: ADMIN_COST_CENTER_NAME, active: true },
+];
+
+export const USERS = [
+  {
+    role_name: 'Autorizador',
+    department_name: 'Finanzas',
+    user_name: 'laura.flores',
+    password: '123',
+    workstation: 'WS104',
+    email: 'laura.flores@empresa.com',
+    phone_number: '555-1004',
+    boss_user_name: null,
+    society_id: 1,
+    supplier: 20000000008,
+  },
+  {
+    role_name: 'Autorizador',
+    department_name: 'Finanzas',
+    user_name: 'diego.hernandez',
+    password: '123',
+    workstation: 'WS105',
+    email: 'diego.hernandez@empresa.com',
+    phone_number: '555-1005',
+    boss_user_name: 'laura.flores',
+    society_id: 1,
+    supplier: 20000000009,
+  },
+  {
+    role_name: 'Solicitante',
+    department_name: 'Finanzas',
+    user_name: 'andres.gomez',
+    password: '123',
+    workstation: 'WS101',
+    email: 'andres.gomez@empresa.com',
+    phone_number: '555-1001',
+    boss_user_name: 'diego.hernandez',
+    society_id: 1,
+    supplier: 20000000013,
+  },
+  {
+    role_name: 'Agencia de viajes',
+    department_name: 'Finanzas',
+    user_name: 'paula.martinez',
+    password: '123',
+    workstation: 'WS102',
+    email: 'paula.martinez@empresa.com',
+    phone_number: '555-1002',
+    boss_user_name: null,
+    society_id: 1,
+    supplier: 20000000010,
+  },
+  {
+    role_name: 'Cuentas por pagar',
+    department_name: 'Finanzas',
+    user_name: 'carlos.ramos',
+    password: '123',
+    workstation: 'WS103',
+    email: 'carlos.ramos@empresa.com',
+    phone_number: '555-1003',
+    boss_user_name: null,
+    society_id: 1,
+    supplier: 20000000011,
+  },
+  {
+    role_name: 'Administrador',
+    department_name: 'Admin',
+    user_name: 'admin',
+    password: '123',
+    workstation: 'WS000',
+    email: 'admin@empresa.com',
+    phone_number: '555-0000',
+    boss_user_name: null,
+    society_id: 1,
+    supplier: 20000000012,
+  },
+  // Tec society group users (society_id: 3 = Tec Servicios)
+  {
+    role_name: 'Administrador',
+    department_name: 'Admin',
+    user_name: 'admin.tec',
+    password: '123',
+    workstation: 'WS000',
+    email: 'admin@tec.com',
+    phone_number: '555-0000',
+    boss_user_name: null,
+    society_id: 3,
+    supplier: 20000000014,
+  },
+  {
+    role_name: 'Autorizador',
+    department_name: 'Finanzas',
+    user_name: 'maria.santos',
+    password: '123',
+    workstation: 'WS204',
+    email: 'maria.santos@tec.com',
+    phone_number: '555-2004',
+    boss_user_name: null,
+    society_id: 3,
+    supplier: 20000000015,
+  },
+  {
+    role_name: 'Autorizador',
+    department_name: 'Finanzas',
+    user_name: 'juan.torres',
+    password: '123',
+    workstation: 'WS205',
+    email: 'juan.torres@tec.com',
+    phone_number: '555-2005',
+    boss_user_name: 'maria.santos',
+    society_id: 3,
+    supplier: 20000000016,
+  },
+  {
+    role_name: 'Solicitante',
+    department_name: 'Finanzas',
+    user_name: 'carlos.silva',
+    password: '123',
+    workstation: 'WS201',
+    email: 'carlos.silva@tec.com',
+    phone_number: '555-2001',
+    boss_user_name: 'juan.torres',
+    society_id: 3,
+    supplier: 20000000017,
+  },
+  {
+    role_name: 'Agencia de viajes',
+    department_name: 'Finanzas',
+    user_name: 'lucia.moreno',
+    password: '123',
+    workstation: 'WS202',
+    email: 'lucia.moreno@tec.com',
+    phone_number: '555-2002',
+    boss_user_name: null,
+    society_id: 3,
+    supplier: 20000000018,
+  },
+  {
+    role_name: 'Cuentas por pagar',
+    department_name: 'Finanzas',
+    user_name: 'fernando.rojas',
+    password: '123',
+    workstation: 'WS203',
+    email: 'fernando.rojas@tec.com',
+    phone_number: '555-2003',
+    boss_user_name: null,
+    society_id: 3,
+    supplier: 20000000019,
+  },
+];
+
+export const ACCOUNTS = [
+  { account_code: '1000', account_name: 'Anticipo', account_type: 'Activo' },
+  { account_code: '1001', account_name: 'Cuenta x pagar Empleado', account_type: 'Pasivo' },
+  { account_code: '1002', account_name: 'Gasto de Viaje', account_type: 'Activo', cost_center_id: 3 },
+  { account_code: '1003', account_name: 'Iva Acreditable', account_type: 'Pasivo' },
+];
+
+export const RECEIPT_TYPE_TO_ACCOUNT = {
+  Hospedaje: '1002',
+  Comida: '1002',
+  Transporte: '1002',
+  Caseta: '1002',
+  Autobús: '1002',
+  Vuelo: '1002',
+  Otro: '1002',
+};
+
+export const SOCIETY_GROUPS = [
+  { id: 1, description: 'Ditta Consulting' },
+  { id: 2, description: 'Tec' },
+]
+
+export const SOCIETIES = [
+  { id: 1, description: 'Ditta Servicios', local_currency: 'MXN', society_group_id: 1 },
+  { id: 2, description: 'Ditta Gestión', local_currency: 'MXN', society_group_id: 1 },
+  { id: 3, description: 'Tec Servicios', local_currency: 'USD', society_group_id: 2 },
+  { id: 4, description: 'Tec Gestión', local_currency: 'USD', society_group_id: 2 },
+]
+
+export const DOCUMENT_TYPES = [
+  {document_id: 'AV', description: 'Anticipo de Viaje'},
+  {document_id: 'GV', description: 'Gasto de Viaje'},
+];
+
+export const RECEIPTS = [
+  {
+    requestIndex: 0,
+    routeIndex: 0,
+    receipt_type_name: 'Hospedaje',
+    amount: 1250.00,
+    currency: 'MXN',
+    exch_rate: 1,
+    refund: true,
+    submission_date: '2025-05-10T14:30:00',
+    validation: 'Aprobado',
+    validation_date: '2025-05-12T10:15:00',
+    description: 'Hospedaje Hotel Majestic Barcelona 3 noches',
+    receipt_number: 'FAC-2025-78412',
+    supplier_name: 'Hotel Majestic SA de CV',
+    pdf_file_name: 'hotel_majestic_78412.pdf',
+    xml_file_name: 'CFDI_Hotel_78412.xml',
+    xml_uuid: 'E5345672-0AF6-4746-87F5-5EFE0409B9D3',
+    xml_rfc_emisor: 'HMA050612P76',
+    xml_rfc_receptor: 'DITT010101000', 
+    xml_nombre_emisor: 'Hotel Majestic SA de CV',
+    xml_fecha: '2025-05-08T18:45:00',
+    xml_total: 1250.00,
+    xml_subtotal: 1033.06,
+    xml_impuestos: 216.94,
+    xml_moneda: 'MXN',
+  },
+  {
+    requestIndex: 1,
+    routeIndex: 0,
+    receipt_type_name: 'Comida',
+    amount: 285.50,
+    currency: 'USD',
+    exch_rate: 17.99,
+    refund: true,
+    submission_date: '2025-05-10T15:00:00',
+    validation: 'Aprobado',
+    validation_date: '2025-05-12T10:20:00',
+    description: 'Comidas durante conferencia en Barcelona',
+    receipt_number: 'TKT-45892',
+    supplier_name: 'Restaurante La Fonda',
+    pdf_file_name: 'alimentos_fonda_45892.pdf',
+    xml_file_name: 'CFDI_Alimentos_45892.xml',
+    xml_uuid: 'A1B2C3D4-E5F6-7890-ABCD-EF1234567890',
+    xml_rfc_emisor: 'FON070415K32',
+    xml_rfc_receptor: 'DITT010101000',
+    xml_nombre_emisor: 'Restaurante La Fonda SA',
+    xml_fecha: '2025-05-09T13:20:00',
+    xml_total: 285.50,
+    xml_subtotal: 235.95,
+    xml_impuestos: 49.55,
+    xml_moneda: 'USD',
+  },
+  {
+    requestIndex: 2,
+    routeIndex: 0,
+    receipt_type_name: 'Transporte',
+    amount: 420.75,
+    currency: 'MXN',
+    exch_rate: 1,
+    refund: true,
+    submission_date: '2025-05-10T15:10:00',
+    validation: 'Aprobado',
+    validation_date: '2025-05-12T10:25:00',
+    description: 'Traslado aeropuerto - hotel Barcelona',
+    receipt_number: 'TAXI-33901',
+    supplier_name: 'Taxi Oficial Barcelona',
+    pdf_file_name: 'taxi_barcelona_33901.pdf',
+    xml_file_name: null,
+    xml_uuid: null,
+    xml_rfc_emisor: null,
+    xml_rfc_receptor: null,
+    xml_nombre_emisor: null,
+    xml_fecha: null,
+    xml_total: 420.75,
+    xml_subtotal: 347.73,
+    xml_impuestos: 73.02,
+    xml_moneda: null,
+  },
+  {
+    requestIndex: 3,
+    routeIndex: null,         
+    receipt_type_name: 'Transporte',
+    amount: 320.00,
+    currency: 'MXN',
+    exch_rate: 1,
+    refund: true,
+    submission_date: '2025-04-16T09:45:00',
+    validation: 'Aprobado',
+    validation_date: '2025-04-18T11:30:00',
+    description: 'Traslados locales CDMX',
+    receipt_number: 'APP-20250415-784',
+    supplier_name: 'Uber México',
+    pdf_file_name: 'uber_20250415.pdf',
+    xml_file_name: 'CFDI_Uber_784.xml',
+    xml_uuid: 'F9A8B7C6-D5E4-3210-9876-543210FEDCBA',
+    xml_rfc_emisor: 'UBE180101ABC',
+    xml_rfc_receptor: 'DITT010101000',
+    xml_nombre_emisor: 'Uber Technologies de México SA de CV',
+    xml_fecha: '2025-04-15T17:30:00',
+    xml_total: 320.00,
+    xml_subtotal: 264.46,
+    xml_impuestos: 55.54,
+    xml_moneda: 'MXN',
+  },
+  {
+    requestIndex: 4,
+    routeIndex: null,
+    receipt_type_name: 'Hospedaje',
+    amount: 890.00,
+    currency: 'MXN',
+    exch_rate: 1,
+    refund: true,
+    submission_date: '2025-04-11T12:00:00',
+    validation: 'Aprobado',
+    validation_date: '2025-04-13T09:00:00',
+    description: 'Noche de hotel por emergencia médica',
+    receipt_number: 'HOSP-2025-11234',
+    supplier_name: 'Hospital HMG',
+    pdf_file_name: 'hospital_hmg_11234.pdf',
+    xml_file_name: 'CFDI_HMG_11234.xml',
+    xml_uuid: '12345678-90AB-CDEF-1234-567890ABCDEF',
+    xml_rfc_emisor: 'HMG050101P89',
+    xml_rfc_receptor: 'DITT010101000',
+    xml_nombre_emisor: 'Hospital HMG SA de CV',
+    xml_fecha: '2025-04-10T20:15:00',
+    xml_total: 890.00,
+    xml_subtotal: 735.54,
+    xml_impuestos: 154.46,
+    xml_moneda: 'MXN',
+  },
+  {
+    requestIndex: 5,
+    routeIndex: 0,
+    receipt_type_name: 'Comida',
+    amount: 420.75,
+    currency: 'USD',
+    exch_rate: 17.98,
+    refund: true,
+    submission_date: '2025-05-13T16:20:00',
+    validation: 'Pendiente',
+    validation_date: null,
+    description: 'Desayuno y comida en congreso internacional',
+    receipt_number: 'FAC-55678',
+    supplier_name: 'Centro de Convenciones Madrid',
+    pdf_file_name: 'alimentos_congreso_55678.pdf',
+    xml_file_name: 'CFDI_Alimentos_55678.xml',
+    xml_uuid: '98765432-1A2B-3C4D-5E6F-7A8B9C0D1E2F',
+    xml_rfc_emisor: 'CCM120101ABC',
+    xml_rfc_receptor: 'DITT010101000',
+    xml_nombre_emisor: 'Centro de Convenciones SA',
+    xml_fecha: '2025-05-12T14:00:00',
+    xml_total: 420.75,
+    xml_subtotal: 347.73,
+    xml_impuestos: 73.02,
+    xml_moneda: 'USD',
+  },
+];
+
+export const REQUESTS = [
+  {
+    notes: 'Solicito viaticos para viaje a conferencia en Barcelona.',
+    requested_fee: 1500.0,
+    imposed_fee: 1500.0,
+    request_days: 3.0,
+    status: 'Finalizado',
+    assigned_to_username: null,
+    document_id: 'GV',
+    authorization_rule_id: 1,
+    exch_rate: 0,
+  },
+  {
+    notes: 'Reembolso por gastos medicos durante viaje.',
+    requested_fee: 0,
+    imposed_fee: 0,
+    request_days: 1.0,
+    status: 'Revisión',
+    assigned_to_username: 'diego.hernandez',
+    document_id: 'GV',
+    authorization_rule_id: 1,
+    exch_rate: 0,
+  },
+  {
+    notes: 'Solicitud de apoyo economico para capacitacion online.',
+    requested_fee: 500.0,
+    imposed_fee: 500.0,
+    request_days: 0.0,
+    status: 'Cotización del Viaje',
+    assigned_to_username: 'carlos.ramos',
+    document_id: 'GV',
+    authorization_rule_id: 1,
+    exch_rate: 0,
+  },
+  {
+    notes: 'Viaticos para taller de liderazgo en Madrid.',
+    requested_fee: 1200.0,
+    imposed_fee: 1200.0,
+    request_days: 2.0,
+    status: 'Atención Agencia de Viajes',
+    assigned_to_username: 'paula.martinez',
+    document_id: 'GV',
+    authorization_rule_id: 1,
+    exch_rate: 0,
+  },
+  {
+    notes: 'Reembolso de transporte.',
+    requested_fee: 300.0,
+    imposed_fee: 250.0,
+    request_days: 0.5,
+    status: 'Comprobación gastos del viaje',
+    assigned_to_username: 'carlos.ramos',
+    document_id: 'GV',
+    authorization_rule_id: 1,
+    exch_rate: 0,
+  },
+  {
+    notes: 'Apoyo para participacion en congreso internacional.',
+    requested_fee: 2000.0,
+    imposed_fee: 1800.0,
+    request_days: 4.0,
+    status: 'Validación de comprobantes',
+    assigned_to_username: 'carlos.ramos',
+    document_id: 'AV',
+    authorization_rule_id: 1,
+    exch_rate: 0,
+  },
+  {
+    notes: 'Gastos operativos extraordinarios.',
+    requested_fee: 0,
+    imposed_fee: 0,
+    request_days: 0.0,
+    status: 'Finalizado',
+    assigned_to_username: null,
+    document_id: 'AV',
+    authorization_rule_id: 1,
+    exch_rate: 0,
+  },
+  {
+    notes: 'Viaje urgente por representacion institucional.',
+    requested_fee: 1750.0,
+    imposed_fee: 1500.0,
+    request_days: 3.5,
+    status: 'Cancelado',
+    assigned_to_username: 'carlos.ramos',
+    document_id: 'AV',
+    authorization_rule_id: 1,
+    exch_rate: 0,
+  },
+  {
+    notes: 'Solicito anticipo para mision tecnica en el extranjero.',
+    requested_fee: 2200.0,
+    imposed_fee: 2000.0,
+    request_days: 5.0,
+    status: 'Finalizado',
+    assigned_to_username: null,
+    document_id: 'AV',
+    authorization_rule_id: 1,
+    exch_rate: 0,
+  }
+];
+
+export const ROUTES = [
+  { fromCountry: 'México', fromCity: 'CDMX', toCountry: 'México', toCity: 'Guadalajara', date: '2025-05-01', begin: '08:00:00', end: '11:00:00', plane_needed: true, hotel_needed: false },
+  { fromCountry: 'México', fromCity: 'Monterrey', toCountry: 'México', toCity: 'Mérida', date: '2025-05-02', begin: '10:30:00', end: '14:30:00', plane_needed: true, hotel_needed: true },
+  { fromCountry: 'México', fromCity: 'Guadalajara', toCountry: 'México', toCity: 'CDMX', date: '2025-05-03', begin: '12:00:00', end: '15:00:00', plane_needed: false, hotel_needed: true },
+  { fromCountry: 'México', fromCity: 'Monterrey', toCountry: 'México', toCity: 'Guadalajara', date: '2025-05-04', begin: '06:00:00', end: '09:00:00', plane_needed: true, hotel_needed: false },
+  { fromCountry: 'México', fromCity: 'CDMX', toCountry: 'Estados Unidos', toCity: 'Nueva York', date: '2025-05-05', begin: '14:00:00', end: '18:00:00', plane_needed: true, hotel_needed: true },
+  { fromCountry: 'Estados Unidos', fromCity: 'Nueva York', toCountry: 'México', toCity: 'CDMX', date: '2025-05-06', begin: '11:00:00', end: '13:00:00', plane_needed: false, hotel_needed: false },
+  { fromCountry: 'México', fromCity: 'CDMX', toCountry: 'España', toCity: 'Madrid', date: '2025-05-07', begin: '09:30:00', end: '12:30:00', plane_needed: true, hotel_needed: false },
+  { fromCountry: 'Reino Unido', fromCity: 'Londres', toCountry: 'Estados Unidos', toCity: 'Los Ángeles', date: '2025-05-08', begin: '15:00:00', end: '18:30:00', plane_needed: true, hotel_needed: true },
+  { fromCountry: 'México', fromCity: 'CDMX', toCountry: 'España', toCity: 'Madrid', date: '2025-05-09', begin: '08:00:00', end: '11:15:00', plane_needed: true, hotel_needed: true },
+];
